+++ src/iconsets/Phosphor.php
@@ -26,7 +26,7 @@ class Phosphor extends RemoteCssIconSet
     // Public Methods
     // =========================================================================
 
-    public function getSettingsHtml(): ?string
+    protected function getVariantSettingsHtml(): ?string
     {
         return Craft::$app->getView()->renderTemplate('icon-picker/icon-sets/phosphor', [
             'iconSet' => $this,